% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{aucC_wrap}
\alias{aucC_wrap}
\title{Quick computation of the AUC}
\usage{
aucC_wrap(label, scores)
}
\arguments{
\item{label}{A factor vector with levels of c("Yes","No") in any order, or a character vector taking the same values.}

\item{scores}{A numeric vector of the same length above. Higher scores must correspond to the "Yes" level of the labels.}
}
\value{
The AUC, defined as the area under the ROC curve computed using \code{label} and \code{scores}.
}
\description{
Quick computation of the AUC
}
\examples{
\dontrun{
y=rbinom(100, size=1, prob=1/2)
label = factor(y, levels=c(0,1), labels=c("No", "Yes"))
scores = runif(100) # higher scores predict label == "Yes"
aucC::aucC_wrap(label=label, scores=scores)
}
}
